seat table연동 오류 수정

+++ api.py
@@ -205,27 +205,28 @@ def reserve():
 # 예약 삭제 
 @board.route("/main", methods=["DELETE"])
 def delete_reserve(): #본인의 예약내용을 삭제할 수 있게 하고, DB에도 삭제한다. 
-    reservationID = request.form['reservationID']
-    seatNum = request.form['seatNum']
-    user_id = request.form['user_id']
-    reserve_data = Reservation.query.filter(Reservation.reservationID == reservationID, Reservation.seatNum == seatNum, Reservation.user_id == user_id).first()
-    seat_data = Seat.query.filter(Seat.seatNum == Reservation.seatNum, Seat.used == 1, Reservation.user_id == Seat.user_id).first()
-    if reserve_data is not None:
-        db.session.delete(reserve_data)
-        db.session.commit()
-        if seat_data is not None:
-            user_id = None
-            used = 0
-            finish_time = None
-
-            seat_data.user_id = user_id
-            seat_data.used = used
-            seat_data.finish_time = finish_time
+    if session.get("login") is not None:
+        reservationID = request.form['reservationID']
+        seatNum = request.form['seatNum']
+        user_id = request.form['user_id']
+        reserve_data = Reservation.query.filter(Reservation.reservationID == reservationID, Reservation.seatNum == seatNum, Reservation.user_id == user_id).first()
+        seat_data = Seat.query.filter(Seat.seatNum == seatNum, Seat.used == 1, Reservation.user_id == Seat.user_id).first()
+        if reserve_data is not None:
+            db.session.delete(reserve_data)
             db.session.commit()
-        
-        return jsonify({'result':'success'})
-    else:
-        return jsonify({'result':'fail'})
+            if seat_data is not None:
+                user_id = None
+                used = 0
+                finish_time = None
+
+                seat_data.user_id = user_id
+                seat_data.used = used
+                seat_data.finish_time = finish_time
+                db.session.commit()
+            
+            return jsonify({'result':'success'})
+        else:
+            return jsonify({'result':'fail'})
 
 # 좌석 초기화 -> 매 오후 12시에 좌석 table을 초기화해주기
 def seat_initialize():
